% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adventureR.R
\docType{package}
\name{adventureR}
\alias{adventureR}
\alias{adventureR-package}
\title{adventureR: the big package of R adventures}
\description{
This package accompanies the \href{http://michaelquinn32.github.io/}{Big Blog of R Adventures}.
Since its only unifying theme is the blog, its primary function is to gather, document and share
the code developed in support of that blog. For that reason, the package should always be considered
"in development."
}
\details{
The code here will also be on the raw side. Instead of running perfectly under most use cases,
I am much more interested in producing code that can be easily interpreted. Again, we're
illustrating how R can solve lots of problems, not how to write the world's most perfect
R code (even if there is such a thing).
}

