% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{f1_score}
\alias{f1_score}
\title{Calculate the F1 score of a predictor}
\usage{
f1_score(prec, rec)
}
\arguments{
\item{prec}{the precision of a predictor}

\item{rec}{the recall of a predictor}
}
\value{
A numeric scalar
}
\description{
Calculate the F1 score of a predictor
}

