% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geyser.R
\name{geyser_simulation}
\alias{autoplot.geyser}
\alias{count_seq}
\alias{count_v}
\alias{eruptions}
\alias{first_eruption}
\alias{geyser_simulation}
\alias{geysers}
\alias{print.summary.geyser}
\alias{summary.geyser}
\title{A Simple Geyser simulation}
\usage{
geyser_simulation(rates, n, timeframe = NULL)

eruptions(rate, timeframe)

geysers(rates, timeframe)

first_eruption(geysers, arrival)

count_v(level, vector)

count_seq(vector, levels)

\method{summary}{geyser}(object, nwindows = 11)

\method{print}{summary.geyser}(object)

## S3 method for class 'geyser'
autoplot(object)
}
\arguments{
\item{rates}{A named \code{list} of eruption rates. The names identify the geysers,
and the values identify the number of hours between eruptions}

\item{n}{The number of iterations in a simulation}

\item{timeframe}{The time window for the eruptions; default is maximum rate}
}
\value{
A \code{geyser} objection containing the following:
 \describe{
     \item{counts}{a data frame tracking which geyser erupts first each simulation}
     \item{frequencies}{a data frame that counts relative rates: count / n simulations}
     \item{first_eruption}{a data frame of the time until each geyser's first eruption}
     \item{n}{the number of simulations}
     \item{rates}{the list of rates passed as an argument}
 }
}
\description{
This function simulates the geyser scenario described in
\href{http://fivethirtyeight.com/tag/the-riddler/}{FiveThirtyEight's Riddler}. The
simulation is built from several support functions. They are all documented together.
}
\details{
The problem states that you arrive at a national park, knowing that three geysers erupt
at fixed intervals. You do not know when these intervals begin. What are the probabilites
that each geyser erupts first? The blog post for this function solved the problem
analytically. This code simulates the scenario to arrive at a solution.
}
\section{Functions}{
\itemize{
\item \code{eruptions}: Simulate the eruption of a single geyser

\item \code{geysers}: Simulate multiple geysers, given rates

\item \code{first_eruption}: Find the time to each geyser's first eruption

\item \code{count_v}: Count the number of times a level appears in a vector

\item \code{count_seq}: Given a vector and a set of levels, count how often
each level appears
}}
\examples{
rates <- list(a = 2, b = 4, c = 6)
geyser_simulation(rates, 10)

}

