% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA}
\alias{LDA}
\alias{quadform}
\title{Linear Discriminant Analysis}
\usage{
LDA(formula, data)

quadform(X, a)
}
\arguments{
\item{formula}{An R formual specifying the LDA}

\item{data}{An R data frame}

\item{X}{a matrix}

\item{a}{a vector of suitable dimension}
}
\value{
A list with the following items
\describe{
     \item{actual}{the original target values}
     \item{x}{the original model data matrix}
     \item{fit}{the class values}
     \item{preds}{the values of the linear combination}
     \item{a}{the LDA coefficients}
     \item{b}{the LDA threshold, or intercept}
}

A numeric scalar
}
\description{
Fit a classifier using a generalization of Fischer's linear discriminant.
}
\details{
The result is found using a relatively simple formula where the components of a discriminant hyper
plane are found separately. This makes it easier to generate predictions later.

\deqn{f(x) = 1 \mathrm{ if } a \cdot x + b > 0}
\deqn{a = \Sigma^{-1} \mu_1 - \mu_0}
\deqn{b = -\frac{1}{2} \mu_1 ' \Sigma^{-1} \mu_1 +
         \frac{1}{2} \mu_0 ' \Sigma^{-1} \mu_0 +
         \log \frac{p_1}{p_0}}
}
\section{Functions}{
\itemize{
\item \code{quadform}: Quickly process quadratic forms (support for LDA)
}}

