% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_bayes.R
\name{naive_bayes}
\alias{log_odds}
\alias{naive_bayes}
\title{Fit a Naive Bayes Classifier}
\usage{
naive_bayes(formula, data, nzv_thresh = 1e-06)

log_odds(x, mus, sigmas)
}
\arguments{
\item{formula}{An R formual specifying the Naive Bayes Classifier}

\item{data}{A data frame to fit the model}

\item{x}{the data}

\item{mus}{the group means (as vectors)}

\item{sigmas}{the group varainces (as vectors)}
}
\value{
A list with the class \code{n} containing the following:
\describe{
     \item{actual}{the original target values}
     \item{x}{the original model data matrix}
     \item{fit}{the class values}
     \item{preds}{the values of the linear combination}
     \item{mus}{a list of the group means, each item is a vector}
     \item{sigmas}{the list of the group variances, each item is a vector}
     \item{p_rat}{the ratio of prior rates}
}

A numeric scalar
}
\description{
Fit a Naive Bayes Classifier
}
\details{
This version of a naive bayes classifier generates predictions through the
log odds of an observation belonging to either class 1 or 0. We can the classifier
"naive" because we calculate the odds assuming that each column is independent.

Our classification rule follows:

\deqn{f(x) = 1 \mathrm{ if } \log \frac{P(Y = 1 | X = x)}{P(Y = 0 | X = x)} }
\deqn{= \log \frac{p_1}{p_0} + \sum_{j = 1}^{p} \left[ \frac{1}{2} \log \frac{\sigma_{0j}^2}{\sigma_{1j}^2} - \frac{(x_j - \mu_{1j})^2}{2 \sigma_{1j}^2} + \frac{(x_j - \mu_{0j})^2}{2 \sigma_{0j}^2}\right]}

The \code{log_odds} function makes it much easier to apply this formula across the entire data frame.
}
\section{Functions}{
\itemize{
\item \code{log_odds}: Calculate Log odds for normally distributed data (a support function)
}}

