% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzv.R
\name{nzv}
\alias{nzv}
\alias{zero_variance}
\title{Remove Zero variance Columns from a data frame}
\usage{
nzv(.data, .target = "target", .thresh = 1e-04)

zero_variance(.col, .thresh)
}
\arguments{
\item{.data}{A data frame}

\item{.thresh}{A minimal theshold for non-zero variance}

\item{.col}{A column of a data frame}
}
\value{
A data frame with the zero variance columns removed
}
\description{
The predicate function \code{zero_variance} checks to see if the variance of a column
is lower than a pre-defined threshold. The function \code{nzv} applies the predicate
to a data frame and drops the columns that fail the test.
}
\section{Functions}{
\itemize{
\item \code{zero_variance}: Predicate function to see if a column is zero variance
}}

