% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lda}
\alias{predict.lda}
\title{Create predictions from an LDA model}
\usage{
\method{predict}{lda}(object, newdata = NULL)
}
\arguments{
\item{object}{An object of class \code{lda}}

\item{newdata}{An optional data frame to use in the prediction}
}
\value{
A list containing the predicted values and classes
}
\description{
Create predictions from an LDA model
}

