% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.nb}
\alias{predict.nb}
\title{Create predictions from a Naive Bayes model}
\usage{
\method{predict}{nb}(object, newdata)
}
\arguments{
\item{object}{An object of class \code{lda}}

\item{newdata}{An optional data frame to use in the prediction}
}
\value{
A list containing the predicted values and classes
}
\description{
Create predictions from a Naive Bayes model
}

