% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{recall}
\alias{recall}
\title{Calculate the recall of a predictor}
\usage{
recall(fit, actual)
}
\arguments{
\item{fit}{the fitted values of a predictor}

\item{the}{original target}
}
\value{
A numeric scalar
}
\description{
Calculate the recall of a predictor
}

