% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_model.R
\name{score_model}
\alias{score_model}
\alias{score_model.lda}
\alias{score_model.nb}
\alias{score_model.train}
\title{Score a model with Precision, Recall and F1}
\usage{
score_model(.object, newdata = NULL, actual = NULL, ...)

\method{score_model}{lda}(.object, newdata = NULL, actual = NULL)

\method{score_model}{nb}(.object, newdata = NULL, actual = NULL)

\method{score_model}{train}(.object, newdata = NULL, actual = NULL)
}
\arguments{
\item{.object}{a model object}

\item{newdata}{an optional new data frame for generating predictions; otherwise score fit on training data}

\item{actual}{an optional vector for the true values; used for assessing predictions on test data}

\item{...}{other parameters passed for scoring}
}
\value{
A list with the Precision, Recall and F1
}
\description{
Score a model with Precision, Recall and F1
}
\section{Methods (by class)}{
\itemize{
\item \code{lda}: Generate scores for lda models

\item \code{nb}: Generate scores for naive bayes models

\item \code{train}: Generate scores for models fit in \code{caret}
}}

