% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Sampling function for bootstrap}
\usage{
split_data(.data, .test_rate = 0.1)
}
\arguments{
\item{.data}{a data frame}

\item{.test_rate}{the (approximate) portion of the data allocated to test data}
}
\value{
a list of two dataframes, the training and test dataset
}
\description{
A function to randomly split a data frame into training and test samples.
}

